# Optimize array pattern matching by caching #deconstruct value