This is follow-up for [the discussion](https://github.com/ruby/ruby/pull/2948#issuecomment-609034219) started in #2948.

## Overview

We can improve array patterns matching performance by keeping the result of the `#deconstruct`  call (or the fact, that the value doesn't respond to it).

The solution proposed in this PR is the following:
- In `compile_case3`:
  - Reserve a stack slot for `#deconstruct` value (with the default value `nil` — haven't tried to deconstruct)
- In `iseq_compile_pattern_each`:
  - Added `deconstructed_pos` argument which is used to track the _distance_ to the reserved slot (since it could be different depending on the node type).

## Benchmark

Compared with the current master ([benchmark](https://github.com/ruby/ruby/commit/ae3ab8ed3a3bb47231eb943d2b4a3168dca771b2)):

```sh
                    first_match
              ./ruby:   1137096.7 i/s 
  /etc/ruby/bin/ruby:   1112537.0 i/s - 1.02x  slower

                     second_match
              ./ruby:    995800.6 i/s 
  /etc/ruby/bin/ruby:    961526.1 i/s - 1.04x  slower

                      third_match
              ./ruby:    895467.2 i/s 
  /etc/ruby/bin/ruby:    822866.8 i/s - 1.09x  slower

                     fourth_match
              ./ruby:    780673.0 i/s 
  /etc/ruby/bin/ruby:    693164.3 i/s - 1.13x  slower
```

## Limitations & future ideas

Currently, only the top-level array patterns are affected. Adding a similar optimization to nested patterns would require additional (and, I think, in most cases useless) stack allocations. Thus, this approach is not optimal, still an ad-hoc optimization.

The proper solution would be to analyze the pattern tree first and generate an optimal _execution path_, which wouldn't have duplicate structural checks. I plan to work on this too, but that would require much more time than the proposed minor improvement. So, I think this patch is still worth to be considered.