@k-tsj 2020-05-16
Great work.

I am positive to merge this pull request.
However, I have been very busy lately, so please give me some time to review.

Could you add some tests?
For example, it seems we need following case.

```ruby
case [0]
in []
in [1] | [0]
  true 
end
```