@palkan 2020-05-18
@k-tsj Added some tests, found some bugs and fixed them 🙂

Unfortunately, the initial idea of including deconstruct instructions only in the first pattern doesn't work well with constant matching, since it's checked first.
So, I rewrote this logic by relying only on the current reserved value on the stack and by adding a couple of new labels to the compiled instructions for branching. The current logic could described as follows:
- Check reserved value:
  - if it's `nil` — we haven't tried to deconstruct yet, jump to `deconstruct`
  - if it's `false` — the matchee doesn't respond to `#deconsturct`, jump to `match_failed`
  - if it's non-falsey — jump to `deconstructed`.

- In the `deconstruct` state:
  - if doesn't respond to `#deconstruct` — put `false` on the stack (instead of `nil`)
  - if does — try to deconstruct and put the result on the stack.