@k-tsj 2020-05-23
I found a bug.
Could you check it?

```
$ ruby -e 'case [[0]]; in [[*a]]; p a;end'
(Expected)
[0]

(Actual)
-e:1:in `<main>': undefined method `length' for 0:Integer (NoMethodError)
```