@palkan 2020-05-23
@k-tsj Fixed (and added a few more tests).