@palkan 2020-06-14
@k-tsj Rebased and extracted this functionality into `iseq_compile_array_deconstruct` to re-use in array and find patterns.